Create Hogwarts_Legacyv2.py
This is the updated code where 16 unique spells (no doubling in spells) can be casted at any given time. Basically you need to assign a F1-F4 for each spell row in Hogwarts Legacy now with this code. for example my first row would be F1. The way that it works is that when you cast a spell, it will automatically cast the correct corresponding spell. 

For example,
If I casted lumos, the raspberry pi will send out the F2 key press (as F2 is assigned to row 2) and then the number 1 will be sent to the game, initating the lumos spell.

then afterwards, I can then cast expelliarmus, the code will then switch the row to F1, send out the number 1 button press and then cast the expellarimus spell.

THIS IS UNTESTED! Please note. The spells used and their gestures may not be 100% accurate. 


Thanks to thistallasian from Reddit for this suggestion.